updated changelog for 0.2.4
it's unlikely that the general protocol (more specifically, the receive checks) will change much in the upcoming versions. Instead, the output will be worked on. Currently, gateway server outputs by printing all communications. in the future all outputs will be put in a json file, carefully formatted to improve readability and usage :)