quite a bit, check commit description
so discord's gateway api changed in a recent build. basically, session settings has less data, session settings is formatted differently, and session settings arrives in 2 parts (READY and READY_SUPPLEMENTAL). On top of fixing that, discum.py was simplified, _Client__gateway_server.runIt now actually returns the targeted response(s) in a list, and Echocage found an api endpoint to check for a user in a guild (getGuildMember)